Move to strcpy in history
Though I want to do everything as safely and efficiently as possible,
the whole **output[1] = '\0'; was causing way too much headache. When
doing this, I was overwriting a local memory reference (my input
string!) a clobbering it. Now, we use the strcpy version so that both
input and output are preserved.